requirements/pro.txt
pydub==0.25.1
python-telegram-bot==20.2.0
openai-whisper==20230314
